+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -17,7 +17,6 @@ import {
 
 import {
   MeetingNotActiveError,
-  createMeetingsError,
   UserInLobbyError,
   NoMediaEstablishedYetError,
   UserNotJoinedError,
@@ -33,7 +32,12 @@ import MeetingStateMachine from './state';
 import {createMuteState} from './muteState';
 import LocusInfo from '../locus-info';
 import Metrics from '../metrics';
-import {trigger, mediaType, error as MetricsError, eventType} from '../metrics/config';
+import {
+  trigger,
+  mediaType as MetricsMediaType,
+  error as MetricsError,
+  eventType,
+} from '../metrics/config';
 import ReconnectionManager from '../reconnection-manager';
 import MeetingRequest from './request';
 import Members from '../members/index';
@@ -90,7 +94,7 @@ import {
   MeetingInfoV2CaptchaError,
 } from '../meeting-info/meeting-info-v2';
 import BrowserDetection from '../common/browser-detection';
-import {ReceiveSlotManager} from '../multistream/receiveSlotManager';
+import {CSI, ReceiveSlotManager} from '../multistream/receiveSlotManager';
 import {MediaRequestManager} from '../multistream/mediaRequestManager';
 import {
   RemoteMediaManager,
@@ -609,7 +613,16 @@ export default class Meeting extends StatelessWebexPlugin {
     /**
      * helper class for managing receive slots (for multistream media connections)
      */
-    this.receiveSlotManager = new ReceiveSlotManager(this);
+    this.receiveSlotManager = new ReceiveSlotManager(
+      (mediaType: MediaType) => {
+        if (!this.mediaProperties?.webrtcMediaConnection) {
+          return Promise.reject(new Error('Webrtc media connection is missing'));
+        }
+
+        return this.mediaProperties.webrtcMediaConnection.createReceiveSlot(mediaType);
+      },
+      (csi: CSI) => (this.members.findMemberByCsi(csi) as any)?.id
+    );
     /**
      * Object containing helper classes for managing media requests for audio/video/screenshare (for multistream media connections)
      * All multistream media requests sent out for this meeting have to go through them.
@@ -2276,6 +2289,8 @@ export default class Meeting extends StatelessWebexPlugin {
           canUnsetMuteOnEntry: ControlsOptionsUtil.canUnsetMuteOnEntry(
             payload.info.userDisplayHints
           ),
+          canSetMuted: ControlsOptionsUtil.canSetMuted(payload.info.userDisplayHints),
+          canUnsetMuted: ControlsOptionsUtil.canUnsetMuted(payload.info.userDisplayHints),
           canStartRecording: RecordingUtil.canUserStart(payload.info.userDisplayHints),
           canStopRecording: RecordingUtil.canUserStop(payload.info.userDisplayHints),
           canPauseRecording: RecordingUtil.canUserPause(payload.info.userDisplayHints),
@@ -2392,6 +2407,30 @@ export default class Meeting extends StatelessWebexPlugin {
         );
       }
     });
+
+    this.locusInfo.on(LOCUSINFO.EVENTS.SELF_REMOTE_VIDEO_MUTE_STATUS_UPDATED, (payload) => {
+      if (payload) {
+        if (this.video) {
+          payload.muted = payload.muted ?? this.video.isRemotelyMuted();
+          payload.unmuteAllowed = payload.unmuteAllowed ?? this.video.isUnmuteAllowed();
+          this.video.handleServerRemoteMuteUpdate(payload.muted, payload.unmuteAllowed);
+        }
+        Trigger.trigger(
+          this,
+          {
+            file: 'meeting/index',
+            function: 'setUpLocusInfoSelfListener',
+          },
+          payload.muted
+            ? EVENT_TRIGGERS.MEETING_SELF_VIDEO_MUTED_BY_OTHERS
+            : EVENT_TRIGGERS.MEETING_SELF_VIDEO_UNMUTED_BY_OTHERS,
+          {
+            payload,
+          }
+        );
+      }
+    });
+
     this.locusInfo.on(LOCUSINFO.EVENTS.SELF_REMOTE_MUTE_STATUS_UPDATED, (payload) => {
       if (payload) {
         if (this.audio) {
@@ -2679,14 +2718,32 @@ export default class Meeting extends StatelessWebexPlugin {
   }
 
   /**
-   * Admit the guest(s) to the call once they are waiting
+   * Admit the guest(s) to the call once they are waiting.
+   * If the host/cohost is in a breakout session, the locus url
+   * of the session must be provided as the authorizingLocusUrl.
+   * Regardless of host/cohost location, the locus Id (lid) in
+   * the path should be the locus Id of the main, which means the
+   * locus url of the api call must be from the main session.
+   * If these loucs urls are not provided, the function will do the check.
    * @param {Array} memberIds
+   * @param {Object} sessionLocusUrls: {authorizingLocusUrl, mainLocusUrl}
    * @returns {Promise} see #members.admitMembers
    * @public
    * @memberof Meeting
    */
-  public admit(memberIds: Array<any>) {
-    return this.members.admitMembers(memberIds);
+  public admit(
+    memberIds: Array<any>,
+    sessionLocusUrls?: {authorizingLocusUrl: string; mainLocusUrl: string}
+  ) {
+    let locusUrls = sessionLocusUrls;
+    if (!locusUrls) {
+      const {locusUrl, mainLocusUrl} = this.breakouts;
+      if (locusUrl && mainLocusUrl) {
+        locusUrls = {authorizingLocusUrl: locusUrl, mainLocusUrl};
+      }
+    }
+
+    return this.members.admitMembers(memberIds, locusUrls);
   }
 
   /**
@@ -3436,7 +3493,7 @@ export default class Meeting extends StatelessWebexPlugin {
           Metrics.postEvent({
             event: eventType.MUTED,
             meeting: this,
-            data: {trigger: trigger.USER_INTERACTION, mediaType: mediaType.AUDIO},
+            data: {trigger: trigger.USER_INTERACTION, mediaType: MetricsMediaType.AUDIO},
           });
         })
         .catch((error) => {
@@ -3489,7 +3546,7 @@ export default class Meeting extends StatelessWebexPlugin {
           Metrics.postEvent({
             event: eventType.UNMUTED,
             meeting: this,
-            data: {trigger: trigger.USER_INTERACTION, mediaType: mediaType.AUDIO},
+            data: {trigger: trigger.USER_INTERACTION, mediaType: MetricsMediaType.AUDIO},
           });
         })
         .catch((error) => {
@@ -3541,7 +3598,7 @@ export default class Meeting extends StatelessWebexPlugin {
           Metrics.postEvent({
             event: eventType.MUTED,
             meeting: this,
-            data: {trigger: trigger.USER_INTERACTION, mediaType: mediaType.VIDEO},
+            data: {trigger: trigger.USER_INTERACTION, mediaType: MetricsMediaType.VIDEO},
           });
         })
         .catch((error) => {
@@ -3593,7 +3650,7 @@ export default class Meeting extends StatelessWebexPlugin {
           Metrics.postEvent({
             event: eventType.UNMUTED,
             meeting: this,
-            data: {trigger: trigger.USER_INTERACTION, mediaType: mediaType.VIDEO},
+            data: {trigger: trigger.USER_INTERACTION, mediaType: MetricsMediaType.VIDEO},
           });
         })
         .catch((error) => {
@@ -6200,6 +6257,27 @@ export default class Meeting extends StatelessWebexPlugin {
    */
   public setDisallowUnmute(enabled: boolean) {
     return this.controlsOptionsManager.setDisallowUnmute(enabled);
+  }
+
+  /**
+   * set the mute all flag for participants if you're the host
+   * @returns {Promise}
+   * @param {boolean} mutedEnabled
+   * @param {boolean} disallowUnmuteEnabled
+   * @param {boolean} muteOnEntryEnabled
+   * @public
+   * @memberof Meeting
+   */
+  public setMuteAll(
+    mutedEnabled: boolean,
+    disallowUnmuteEnabled: boolean,
+    muteOnEntryEnabled: boolean
+  ) {
+    return this.controlsOptionsManager.setMuteAll(
+      mutedEnabled,
+      disallowUnmuteEnabled,
+      muteOnEntryEnabled
+    );
   }
 
   /**