+++ packages/@webex/plugin-meetings/src/members/index.ts
@@ -733,15 +733,22 @@ export default class Members extends StatelessWebexPlugin {
   /**
    * Admits waiting members (invited guests to meeting)
    * @param {Array} memberIds
+   * @param {Object} sessionLocusUrls: {authorizingLocusUrl, mainLocusUrl}
    * @returns {Promise}
    * @public
    * @memberof Members
    */
-  public admitMembers(memberIds: Array<any>) {
+  public admitMembers(
+    memberIds: Array<any>,
+    sessionLocusUrls?: {authorizingLocusUrl: string; mainLocusUrl: string}
+  ) {
     if (isEmpty(memberIds)) {
       return Promise.reject(new ParameterError('No member ids provided to admit.'));
     }
-    const options = MembersUtil.generateAdmitMemberOptions(memberIds, this.locusUrl);
+    const options = {
+      sessionLocusUrls,
+      ...MembersUtil.generateAdmitMemberOptions(memberIds, this.locusUrl),
+    };
 
     return this.membersRequest.admitMember(options);
   }
@@ -773,11 +780,12 @@ export default class Members extends StatelessWebexPlugin {
    * Audio mutes another member in a meeting
    * @param {String} memberId
    * @param {boolean} [mute] default true
+   * @param {boolean} [isAudio] default true
    * @returns {Promise}
    * @public
    * @memberof Members
    */
-  public muteMember(memberId: string, mute = true) {
+  public muteMember(memberId: string, mute = true, isAudio = true) {
     if (!this.locusUrl) {
       return Promise.reject(
         new ParameterError(
@@ -790,7 +798,7 @@ export default class Members extends StatelessWebexPlugin {
         new ParameterError('The member id must be defined to mute the member.')
       );
     }
-    const options = MembersUtil.generateMuteMemberOptions(memberId, mute, this.locusUrl);
+    const options = MembersUtil.generateMuteMemberOptions(memberId, mute, this.locusUrl, isAudio);
 
     return this.membersRequest.muteMember(options);
   }