+++ CHANGELOG.md
@@ -2,6 +2,28 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.44](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.43...v3.0.0-beta.44) (2023-03-15)
+
+
+### Features
+
+* Host/Cohost - Admit user in lobby to main meeting from Breakout Session SPARK-400367 ([#2658](https://github.com/webex/webex-js-sdk/issues/2658)) ([c43e6b3](https://github.com/webex/webex-js-sdk/commit/c43e6b36212608e4f0d3f1edbb3467382e98aafa))
+* **remote-video-mute:** respect remote video mute ([#2665](https://github.com/webex/webex-js-sdk/issues/2665)) ([e6b684b](https://github.com/webex/webex-js-sdk/commit/e6b684b2b25f30360e9152f290fe6bd474066508))
+
+## [3.0.0-beta.43](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.42...v3.0.0-beta.43) (2023-03-14)
+
+
+### Features
+
+* **meetings:** mute all ([#2667](https://github.com/webex/webex-js-sdk/issues/2667)) ([c198945](https://github.com/webex/webex-js-sdk/commit/c19894556bab55b5707638a607e350a41a1ffdb5))
+
+## [3.0.0-beta.42](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.41...v3.0.0-beta.42) (2023-03-13)
+
+
+### Bug Fixes
+
+* remove dependency of ReceiveSlotManager on Meeting ([#2664](https://github.com/webex/webex-js-sdk/issues/2664)) ([82a65e2](https://github.com/webex/webex-js-sdk/commit/82a65e2eb6ce46b3d88e7784d6c60017c7c986b3))
+
 ## [3.0.0-beta.41](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.40...v3.0.0-beta.41) (2023-03-09)